finish migrating helper tests
